LayoutTests/fast/forms/week-multiple-fields/week-multiple-fields-mouse-events-expected.txt

Multiple fields UI of date input type with mouse events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Please run this with DRT or WTR.

Test following mouse actions:
Mouse click to focus each of sub-fields
Mouse click on the spin button to update each of sub-fields

==> Focus on the week field.
PASS input.value is "2345-W08"
PASS input.value is "2345-W09"
PASS input.value is "2345-W08"
PASS window.getSelection().rangeCount is 0

==> Focus on the year field.
PASS input.value is "2346-W08"
PASS input.value is "2347-W08"
PASS input.value is "2346-W08"
PASS window.getSelection().rangeCount is 0

==> Click on a disabled field.
PASS input.value is "2346-W08"

==> Click on a read-only field.
PASS input.value is "2346-W08"

==> Click on clear button.
PASS input.value is "2346-W08"
PASS input.value is "2346-W08"
PASS input.value is ""

PASS successfullyParsed is true

TEST COMPLETE
